pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-news"
version = "0.1.0"
description = "A personal morning news digest: RSS fetcher, deduper, scorer, summarizer, TTS, and a private podcast feed."
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.40.0",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "PyYAML>=6.0.2",
    "rapidfuzz>=3.9.0",
    "Jinja2>=3.1.4",
    "feedgen>=1.0.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0"]

[project.scripts]
daily-news = "daily_news.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
